+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/main/java/com/tencent/cloud/polaris/circuitbreaker/common/PolarisResultToErrorCode.java
@@ -17,6 +17,7 @@
 
 package com.tencent.cloud.polaris.circuitbreaker.common;
 
+import com.tencent.cloud.polaris.context.CircuitBreakerStatusCodeException;
 import com.tencent.polaris.circuitbreak.api.pojo.ResultToErrorCode;
 import feign.FeignException;
 
@@ -48,6 +49,9 @@ public class PolarisResultToErrorCode implements ResultToErrorCode {
 		else if (checkClassExist("org.springframework.web.reactive.function.client.WebClientResponseException")
 				&& e instanceof WebClientResponseException) {
 			return ((WebClientResponseException) e).getRawStatusCode();
+		}
+		else if (e instanceof CircuitBreakerStatusCodeException) {
+			return ((CircuitBreakerStatusCodeException) e).getRawStatusCode();
 		}
 		return -1;
 	}