+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/config/PolarisConfigProperties.java
@@ -69,7 +69,7 @@ public class PolarisConfigProperties {
 	/**
 	 * Attribute refresh type.
 	 */
-	private RefreshType refreshType = RefreshType.REFLECT;
+	private RefreshType refreshType = RefreshType.REFRESH_CONTEXT;
 
 	/**
 	 * List of injected configuration files.
@@ -94,6 +94,11 @@ public class PolarisConfigProperties {
 	 * If internal config file enabled.
 	 */
 	private boolean internalEnabled = true;
+
+	/**
+	 * if check address is enabled.
+	 */
+	private boolean checkAddress = true;
 
 	public boolean isEnabled() {
 		return enabled;
@@ -191,6 +196,14 @@ public class PolarisConfigProperties {
 		this.internalEnabled = internalEnabled;
 	}
 
+	public boolean isCheckAddress() {
+		return checkAddress;
+	}
+
+	public void setCheckAddress(boolean checkAddress) {
+		this.checkAddress = checkAddress;
+	}
+
 	@Override
 	public String toString() {
 		return "PolarisConfigProperties{" +
@@ -206,6 +219,7 @@ public class PolarisConfigProperties {
 				", dataSource='" + dataSource + '\'' +
 				", localFileRootPath='" + localFileRootPath + '\'' +
 				", internalEnabled=" + internalEnabled +
+				", checkAddress=" + checkAddress +
 				'}';
 	}
 }