+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/instrument/scg/EnhancedGatewayGlobalFilter.java
@@ -18,6 +18,8 @@
 package com.tencent.cloud.rpc.enhancement.instrument.scg;
 
 import java.net.URI;
+import java.nio.charset.StandardCharsets;
+import java.util.Optional;
 
 import com.tencent.cloud.common.constant.OrderConstant;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginContext;
@@ -25,15 +27,24 @@ import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginRunner;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedRequestContext;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedResponseContext;
+import com.tencent.polaris.api.utils.CollectionUtils;
+import com.tencent.polaris.api.utils.StringUtils;
+import com.tencent.polaris.circuitbreak.client.exception.CallAbortedException;
 import reactor.core.publisher.Mono;
 
 import org.springframework.cloud.client.DefaultServiceInstance;
+import org.springframework.cloud.client.ServiceInstance;
+import org.springframework.cloud.client.loadbalancer.Response;
 import org.springframework.cloud.gateway.filter.GatewayFilterChain;
 import org.springframework.cloud.gateway.filter.GlobalFilter;
 import org.springframework.cloud.gateway.route.Route;
 import org.springframework.core.Ordered;
+import org.springframework.core.io.buffer.DataBuffer;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.server.reactive.ServerHttpResponse;
 import org.springframework.web.server.ServerWebExchange;
 
+import static org.springframework.cloud.gateway.support.ServerWebExchangeUtils.GATEWAY_LOADBALANCER_RESPONSE_ATTR;
 import static org.springframework.cloud.gateway.support.ServerWebExchangeUtils.GATEWAY_REQUEST_URL_ATTR;
 import static org.springframework.cloud.gateway.support.ServerWebExchangeUtils.GATEWAY_ROUTE_ATTR;
 
@@ -52,18 +63,40 @@ public class EnhancedGatewayGlobalFilter implements GlobalFilter, Ordered {
 
 	@Override
 	public Mono<Void> filter(ServerWebExchange originExchange, GatewayFilterChain chain) {
+		Response<ServiceInstance> serviceInstanceResponse = originExchange.getAttribute(GATEWAY_LOADBALANCER_RESPONSE_ATTR);
+		String serviceId = Optional.ofNullable(serviceInstanceResponse).map(Response::getServer).
+				map(ServiceInstance::getServiceId).orElse(null);
+
 		EnhancedPluginContext enhancedPluginContext = new EnhancedPluginContext();
 
 		EnhancedRequestContext enhancedRequestContext = EnhancedRequestContext.builder()
 				.httpHeaders(originExchange.getRequest().getHeaders())
 				.httpMethod(originExchange.getRequest().getMethod())
 				.url(originExchange.getRequest().getURI())
+				.serviceUrl(getServiceUri(originExchange, serviceId))
 				.build();
 		enhancedPluginContext.setRequest(enhancedRequestContext);
 		enhancedPluginContext.setOriginRequest(originExchange);
 
 		// Run pre enhanced plugins.
-		pluginRunner.run(EnhancedPluginType.Client.PRE, enhancedPluginContext);
+		try {
+			pluginRunner.run(EnhancedPluginType.Client.PRE, enhancedPluginContext);
+		}
+		catch (CallAbortedException e) {
+			if (e.getFallbackInfo() == null) {
+				throw e;
+			}
+			// circuit breaker fallback, not need to run post/exception enhanced plugins.
+			ServerHttpResponse response = originExchange.getResponse();
+			HttpStatus httpStatus = HttpStatus.resolve(e.getFallbackInfo().getCode());
+			response.setStatusCode(httpStatus != null ? httpStatus : HttpStatus.INTERNAL_SERVER_ERROR);
+			if (CollectionUtils.isNotEmpty(e.getFallbackInfo().getHeaders())) {
+				e.getFallbackInfo().getHeaders().forEach(response.getHeaders()::set);
+			}
+			String body = Optional.of(e.getFallbackInfo().getBody()).orElse("");
+			DataBuffer dataBuffer = response.bufferFactory().wrap(body.getBytes(StandardCharsets.UTF_8));
+			return response.writeWith(Mono.just(dataBuffer));
+		}
 		// Exchange may be changed in plugin
 		ServerWebExchange exchange = (ServerWebExchange) enhancedPluginContext.getOriginRequest();
 		long startTime = System.currentTimeMillis();
@@ -73,9 +106,9 @@ public class EnhancedGatewayGlobalFilter implements GlobalFilter, Ordered {
 					URI uri = exchange.getAttribute(GATEWAY_REQUEST_URL_ATTR);
 					enhancedPluginContext.getRequest().setUrl(uri);
 					if (uri != null) {
-						if (route != null && route.getUri().getScheme().contains("lb")) {
+						if (route != null && route.getUri().getScheme().contains("lb") && StringUtils.isNotEmpty(serviceId)) {
 							DefaultServiceInstance serviceInstance = new DefaultServiceInstance();
-							serviceInstance.setServiceId(route.getUri().getHost());
+							serviceInstance.setServiceId(serviceId);
 							serviceInstance.setHost(uri.getHost());
 							serviceInstance.setPort(uri.getPort());
 							enhancedPluginContext.setTargetServiceInstance(serviceInstance, null);
@@ -112,5 +145,20 @@ public class EnhancedGatewayGlobalFilter implements GlobalFilter, Ordered {
 	@Override
 	public int getOrder() {
 		return OrderConstant.Client.Scg.ENHANCED_FILTER_ORDER;
+	}
+
+	private URI getServiceUri(ServerWebExchange originExchange, String serviceId) {
+		URI uri = originExchange.getAttribute(GATEWAY_REQUEST_URL_ATTR);
+		if (StringUtils.isEmpty(serviceId) || uri == null) {
+			return null;
+		}
+		try {
+			return new URI(originExchange.getRequest().getURI().getScheme(),
+					serviceId, uri.getPath(),
+					originExchange.getRequest().getURI().getRawQuery());
+		}
+		catch (Exception e) {
+			return null;
+		}
 	}
 }