+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Change Log
 ---
 
+- [feat: implement circuit breaker in enhance plugin, support listen config group, support refresh single config in refresh_context mode.](https://github.com/Tencent/spring-cloud-tencent/pull/1501)